% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addCountMatrices}
\alias{addCountMatrices}
\title{Load count matrices}
\usage{
addCountMatrices(version, samples, data_path, n.cores, verbose)
}
\description{
Load count matrices
}
\keyword{internal}
