% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addDetailedMetricsInner}
\alias{addDetailedMetricsInner}
\title{Add detailed metrics (= total UMI count and gene count per cell).}
\usage{
addDetailedMetricsInner(cms, verbose = TRUE, n.cores = 1)
}
\arguments{
\item{cms}{List containing the count matrices.}

\item{verbose}{Print messages (default = TRUE).}

\item{n.cores}{Number of cores for the calculations (default = 1).}
}
\value{
data frame
}
\description{
Add detailed metrics, uses not preprocessed count matrices.
}
\keyword{internal}
