% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addPlotStats}
\alias{addPlotStats}
\title{Add statistics to plot}
\usage{
addPlotStats(
  p,
  comp.group,
  metadata,
  h.adj = 0.05,
  primary.test,
  secondary.test,
  exact = FALSE
)
}
\arguments{
\item{p}{Plot to add statistics to.}

\item{comp.group}{Comparison metric.}

\item{metadata}{Metadata for samples.}

\item{h.adj}{Position of statistics test p value as \% of max(y) (default = 0.05).}

\item{primary.test}{Primary statistical test, e.g. "anova", "kruskal.test".}

\item{secondary.test}{Secondary statistical test, e.g. "t-test", "wilcox.test"}

\item{exact}{Whether to display exact p-values instead of rounded values (default = FALSE).}
}
\value{
ggplot2 object
}
\description{
Use ggpubr to add statistics to plots.
}
\keyword{internal}
