% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addPlotStatsSamples}
\alias{addPlotStatsSamples}
\title{Add statistics to plot}
\usage{
addPlotStatsSamples(
  p,
  comp.group,
  metadata,
  h.adj = 0.05,
  exact = FALSE,
  second.comp.group
)
}
\arguments{
\item{p}{Plot to add statistics to.}

\item{comp.group}{Metadata column used as the primary grouping variable. Must match a column name of metadata (default = self$comp.group).}

\item{metadata}{Metadata for samples.}

\item{h.adj}{Relative vertical offset for displaying the statistical test p-value in the plot, expressed as a \% of the maximum y-value (default = 0.05).}

\item{exact}{Whether to display exact p-values instead of rounded values (default = FALSE).}

\item{second.comp.group}{Second comparison metric.}
}
\value{
ggplot2 object
}
\description{
Use ggpubr to add statistics to samples or plot
}
\keyword{internal}
