% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{checkDataPath}
\alias{checkDataPath}
\title{Check data path}
\usage{
checkDataPath(data.path)
}
\arguments{
\item{data.path}{character Path to be checked}
}
\description{
Helper function to check that data.path is not NULL
}
\keyword{internal}
