% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{checkPackageInstalled}
\alias{checkPackageInstalled}
\title{Check whether a package is installed}
\usage{
checkPackageInstalled(
  pkgs,
  details = "to run this function",
  install.help = NULL,
  bioc = FALSE,
  cran = FALSE
)
}
\description{
This function is shamelessly stolen from github.com/kharchenkolab/Cacoa
}
\keyword{internal}
