% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{createUniqueCellNames}
\alias{createUniqueCellNames}
\title{Create unique cell names}
\usage{
createUniqueCellNames(cms, sample.names, sep = "!!")
}
\arguments{
\item{cms}{list List of count matrices, should be named (optional)}

\item{sample.names}{character Optional, list of sample names}

\item{sep}{character Separator between sample IDs and cell IDs (default = "!!")}
}
\description{
Create unique cell names from sample IDs and cell IDs
}
\keyword{internal}
