% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{filterVector}
\alias{filterVector}
\title{Create filtering vector}
\usage{
filterVector(num.vec, name, filter, samples, sep = "!!")
}
\arguments{
\item{num.vec}{numeric Numeric vector to create filter on}

\item{name}{character Name of filter}

\item{filter}{numeric Either a single numeric value or a numeric value with length of samples}

\item{samples}{character Sample IDs}

\item{sep}{character Separator to split cells by into sample-wise lists (default = "!!")}
}
\description{
Create logical filtering vector based on a numeric vector and a (sample-wise) cutoff
}
\examples{
\dontrun{
filter <- filterVector(num.vec = 1:10 \%>\% setNames(c(rep("sample1!!cell",5),rep("sample2!!cell",5))), 
name = "test", 
filter = c(3,7) \%>\% setNames(c("sample1","sample2")), 
samples = c("sample1","sample2"))
}
}
\keyword{internal}
