% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{labelsFilter}
\alias{labelsFilter}
\title{Get labels for percentage of filtered cells}
\usage{
labelsFilter(filter.data)
}
\arguments{
\item{filter.data}{Data frame containing the mitochondrial fraction, depth and doublets per sample}
}
\description{
Get labels for percentage of filtered cells
}
\keyword{internal}
