% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{loadCountMatrices}
\alias{loadCountMatrices}
\title{Load count matrices}
\usage{
loadCountMatrices(
  data_path,
  samples = NULL,
  transcript = c("SYMBOL", "ENSEMBL"),
  sep = "!!",
  n.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data_path}{Path to cellranger count data.}

\item{samples}{Vector of samples.}

\item{transcript}{The type of transcript, SYMBOL or ENSEMBLE (default = "SYMBOL").}

\item{sep}{Separator for cell names (default = "!!").}

\item{n.cores}{Number of cores for the calculations (default = 1).}

\item{verbose}{Print messages (default = TRUE).}
}
\value{
data frame
}
\description{
Load gene expression count data
}
\examples{
cms <- loadCountMatrices(data_path = crm$data_path, samples = crm$metadata$samples, transcript = "SYMBOL", n.cores = crm$n.cores)
}
\keyword{internal}
