% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{percFilter}
\alias{percFilter}
\title{Calculate percentage of filtered cells}
\usage{
percFilter(filter.data, filter = "mito", no.vars = 1)
}
\arguments{
\item{filter.data}{Data frame containing the mitochondrial fraction, depth and doublets per sample.}

\item{filter}{The variable to filter (default = "mito")}

\item{no.vars}{numeric Number of variables (default = 1)}
}
\value{
vector
}
\description{
Calculate percentage of filtered cells based on the filter
}
\keyword{internal}
