% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{plotGeom}
\alias{plotGeom}
\title{Plot the data as points, as bars as a histogram, or as a violin}
\usage{
plotGeom(g, plot.geom, col, pal = NULL)
}
\arguments{
\item{g}{ggplot2 object}

\item{plot.geom}{The plot.geom to use, "point", "bar", "histogram", or "violin".}

\item{pal}{character Palette (default = NULL)}
}
\value{
geom
}
\description{
Plot the data as points, barplot, histogram or violin
}
\keyword{internal}
