% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{readParse}
\alias{readParse}
\title{Load Parse count matrices}
\usage{
readParse(
  data.path,
  samples = NULL,
  raw = FALSE,
  sep = "!!",
  unique.names = TRUE,
  n.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data.path}{Path to Parse count data.}

\item{samples}{Vector of sample names (default = NULL)}

\item{raw}{logical Add raw count matrices (default = FALSE)}

\item{sep}{Separator for cell names (default = "!!").}

\item{n.cores}{Number of cores for the calculations (default = 1).}

\item{verbose}{Print messages (default = TRUE).}
}
\value{
data frame
}
\description{
Load gene expression count data
}
\keyword{internal}
