% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addDetailedMetricsInner}
\alias{addDetailedMetricsInner}
\title{Add detailed metrics}
\usage{
addDetailedMetricsInner(cm.list, verbose = TRUE, n.cores = 1)
}
\description{
Add detailed metrics, requires to load raw count matrices using pagoda2
}
\keyword{internal}
