% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addPlotStats}
\alias{addPlotStats}
\title{Add statistics to plot}
\usage{
addPlotStats(p, comp_group, metadata, pos = 0, exact = FALSE)
}
\description{
Use ggpubr to add statistics to plots
}
\keyword{internal}
