% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addPlotStats}
\alias{addPlotStats}
\title{Add statistics to plot}
\usage{
addPlotStats(p, comp_group, metadata, h.adj = 0.05, exact = FALSE)
}
\description{
Use ggpubr to add statistics to plots
}
\keyword{internal}
