% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addPlotStatsSamples}
\alias{addPlotStatsSamples}
\title{Add statistics to plot}
\usage{
addPlotStatsSamples(
  p,
  comp_group,
  metadata,
  h.adj = 0.05,
  exact = FALSE,
  second_comp_group
)
}
\arguments{
\item{p}{Plot to add statistics to.}

\item{comp_group}{Comparison metric.}

\item{metadata}{Metadata for samples.}

\item{h.adj}{Position of statistics test p value as \% of max(y) (default = 0.05).}

\item{exact}{Whether to calculate exact p values (default = FALSE).}

\item{second_comp_group}{Second comparison metric.}
}
\value{
ggplot2 object
}
\description{
Use ggpubr to add statistics to samples ar plot
}
\examples{
addPlotStats(p, comp_group = "sex", metadata = crm$metadata, second_comp_group = "condition")
}
\keyword{internal}
