% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addSummaryMetrics}
\alias{addSummaryMetrics}
\title{Add summary metrics}
\usage{
addSummaryMetrics(data_path, metadata, verbose = TRUE)
}
\description{
Add summary metrics by reading Cell Ranger metrics summary files
}
\keyword{internal}
