% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{addSummaryMetrics}
\alias{addSummaryMetrics}
\title{Add summary metrics}
\usage{
addSummaryMetrics(data_path, metadata, n.cores = 1, verbose = TRUE)
}
\arguments{
\item{data_path}{Path to cellranger count data.}

\item{metadata}{Metadata for samples.}

\item{n.cores}{Number of cores for the calculations (default = 1).}

\item{verbose}{Print messages (default = TRUE).}
}
\value{
data frame
}
\description{
Add summary metrics by reading Cell Ranger metrics summary files.
}
\examples{
summary.metrics <- addSummaryMetrics(data_path = crm$data_path, metadata = crm$metadata, n.cores = crm$n.cores)
}
\keyword{internal}
