% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{checkCompGroup}
\alias{checkCompGroup}
\title{Set correct 'comp_group' parameter}
\usage{
checkCompGroup(comp_group, category, verbose = TRUE)
}
\arguments{
\item{comp_group}{Comparison metric}

\item{category}{Comparison metric to use if comp_group is not provided}

\item{verbose}{Print messages (default = TRUE)}
}
\description{
Set comp_group to 'category' if null
}
\keyword{internal}
