% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{checkCompMeta}
\alias{checkCompMeta}
\title{Check whether 'comp_group' is in metadata}
\usage{
checkCompMeta(comp_group, metadata)
}
\arguments{
\item{comp_group}{Comparison metric.}

\item{metadata}{Metadata for samples.}
}
\value{
nothing or stop
}
\description{
Checks whether 'comp_group' is any of the column names in metadata.
}
\examples{
checkCompMeta(comp_group = "sex", metadata = crm$metadata)
}
\keyword{internal}
