% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{checkCompMeta}
\alias{checkCompMeta}
\title{Check whether 'comp.group' is in metadata}
\usage{
checkCompMeta(comp.group, metadata)
}
\arguments{
\item{comp.group}{Comparison metric.}

\item{metadata}{Metadata for samples.}
}
\value{
nothing or stop
}
\description{
Checks whether 'comp.group' is any of the column names in metadata.
}
\keyword{internal}
