% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{checkCompMeta}
\alias{checkCompMeta}
\title{Check whether 'comp_group' is in metadata}
\usage{
checkCompMeta(comp_group, metadata)
}
\description{
Checks whether 'comp_group' is any of the column names in metadata
}
\keyword{internal}
