% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkDetailedMetrics}
\alias{checkDetailedMetrics}
\title{Add detailed metrics if they don't exist}
\usage{
checkDetailedMetrics(detailed_metrics, verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages (default = TRUE)}
}
\description{
Internal function for adding detailed metrics
}
\keyword{internal}
