% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRMetrics.R
\name{CRMetrics}
\alias{CRMetrics}
\title{CRMetrics class object}
\description{
CRMetrics class object

CRMetrics class object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{(default = NULL)}

\item{\code{data_path}}{(default = NULL)}

\item{\code{summary_metrics}}{(default = NULL)}

\item{\code{detailed_metrics}}{(default = NULL)}

\item{\code{comp_group}}{(default = NULL)}

\item{\code{verbose}}{(default = TRUE)}

\item{\code{pal}}{(default = NULL)}

\item{\code{theme}}{(default = NULL)}

\item{\code{n.cores}}{(default = 1)
Inner function to add detailed metrics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CRMetrics$new()}}
\item \href{#method-addDetailedMetrics}{\code{CRMetrics$addDetailedMetrics()}}
\item \href{#method-addComparison}{\code{CRMetrics$addComparison()}}
\item \href{#method-plotSamples}{\code{CRMetrics$plotSamples()}}
\item \href{#method-plotGeneCounts}{\code{CRMetrics$plotGeneCounts()}}
\item \href{#method-plotUmiCounts}{\code{CRMetrics$plotUmiCounts()}}
\item \href{#method-plotMedianUmi}{\code{CRMetrics$plotMedianUmi()}}
\item \href{#method-plotSummaryStats}{\code{CRMetrics$plotSummaryStats()}}
\item \href{#method-plotMedianGene}{\code{CRMetrics$plotMedianGene()}}
\item \href{#method-plotCells}{\code{CRMetrics$plotCells()}}
\item \href{#method-plotUMAP}{\code{CRMetrics$plotUMAP()}}
\item \href{#method-plotDepth}{\code{CRMetrics$plotDepth()}}
\item \href{#method-plotUMAPDepth}{\code{CRMetrics$plotUMAPDepth()}}
\item \href{#method-saveSummaryMetrics}{\code{CRMetrics$saveSummaryMetrics()}}
\item \href{#method-saveDetailedMetrics}{\code{CRMetrics$saveDetailedMetrics()}}
\item \href{#method-clone}{\code{CRMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$new(
  data_path,
  metadata_file = NULL,
  comp_group = NULL,
  detailed_metrics = FALSE,
  verbose = TRUE,
  pal = "pearl_earring",
  theme = theme_bw(),
  n.cores = 1,
  version = "V2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Print messages (default = TRUE)}

\item{\code{pal}}{Palette from package 'dutchmasters' for plotting (default: "pearl_earring")}

\item{\code{theme}}{Ggplot2 theme (default: theme_bw())}

\item{\code{n.cores}}{Number of cores for calculations (default = 1)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-addDetailedMetrics}{}}}
\subsection{Method \code{addDetailedMetrics()}}{
Function to read in detailed metrics. This is not done upon initialization for speed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addDetailedMetrics(
  version = "V2",
  samples = self$metadata$sample,
  data_path = self$data_path,
  n.cores = self$n.cores,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{(default = "V2)}

\item{\code{samples}}{vector containing samples. Default is to extract this information from self$metadata$sample}

\item{\code{data_path}}{Path to data (default = self$data_path)}

\item{\code{n.cores}}{Number of cores for the calculations (default = self$n.cores)}

\item{\code{verbose}}{Print messages or not (default = self$verbose)
Add comparison group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addComparison"></a>}}
\if{latex}{\out{\hypertarget{method-addComparison}{}}}
\subsection{Method \code{addComparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addComparison(comp_group, metadata = self$metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric, e.g. "group", "samples"}

\item{\code{metadata}}{Metadata for samples (default = self$metadata)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector
Plot samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSamples"></a>}}
\if{latex}{\out{\hypertarget{method-plotSamples}{}}}
\subsection{Method \code{plotSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSamples(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  second_comp_group = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric (default = self$comp_group)}

\item{\code{h.adj}}{Position of statistics test p value as \% of max(y) (default = 0.05)}

\item{\code{exact}}{Whether to calculate exact p values (default = FALSE)}

\item{\code{metadata}}{Metadata for samples (default = self$metadata)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
Plot gene counts of raw data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneCounts"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneCounts}{}}}
\subsection{Method \code{plotGeneCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotGeneCounts(
  comp_group = self$comp_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Plot UMI counts of raw data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotUmiCounts"></a>}}
\if{latex}{\out{\hypertarget{method-plotUmiCounts}{}}}
\subsection{Method \code{plotUmiCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUmiCounts(
  comp_group = self$comp_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Plot median UMI counts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMedianUmi"></a>}}
\if{latex}{\out{\hypertarget{method-plotMedianUmi}{}}}
\subsection{Method \code{plotMedianUmi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotMedianUmi(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Plot all summary stats or a selected list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSummaryStats"></a>}}
\if{latex}{\out{\hypertarget{method-plotSummaryStats}{}}}
\subsection{Method \code{plotSummaryStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSummaryStats(
  comp_group = self$comp_group,
  metrics = NULL,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Summary stat plot for median gene number per cell}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMedianGene"></a>}}
\if{latex}{\out{\hypertarget{method-plotMedianGene}{}}}
\subsection{Method \code{plotMedianGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotMedianGene(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Plot cells}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCells"></a>}}
\if{latex}{\out{\hypertarget{method-plotCells}{}}}
\subsection{Method \code{plotCells()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotCells(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Plot cells in a UMAP using Conos and color by depth and doublets}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotUMAP"></a>}}
\if{latex}{\out{\hypertarget{method-plotUMAP}{}}}
\subsection{Method \code{plotUMAP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUMAP(
  cms = self$cm.list,
  preprocess = "Pagoda2",
  n.cores = self$n.cores,
  verbose = self$verbose,
  force = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cms}}{The count matrices (default = self$cm.list)}

\item{\code{preprocess}}{The method used for preprocessing, either Pagoda2 or Seurat (default = "Pagoda2")}

\item{\code{n.cores}}{The number of cores to use (default = self$n.cores)}

\item{\code{verbose}}{boolean Print progress (default = self$verbose)}

\item{\code{force}}{boolean Force calculations (default = FALSE)
Plot the depth in histogram}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDepth"></a>}}
\if{latex}{\out{\hypertarget{method-plotDepth}{}}}
\subsection{Method \code{plotDepth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotDepth(cutoff_depth = 10000, per_sample = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cutoff_depth}}{The depth cutoff to color the UMAP (default = 1e4)}

\item{\code{per_sample}}{Whether to plot the depth per sample or for all the samples (default = FALSE)
Plot UMAP with coloring for the depth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotUMAPDepth"></a>}}
\if{latex}{\out{\hypertarget{method-plotUMAPDepth}{}}}
\subsection{Method \code{plotUMAPDepth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUMAPDepth(
  depth = self$depth,
  cutoff_depth = 1000,
  per_sample = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{The object which holds the depth (default = self$depth)}

\item{\code{cutoff_depth}}{The depth cutoff to color the UMAP (default = 1e4)}

\item{\code{per_sample}}{Whether to plot the depth per sample or for all the samples (default = FALSE)}

\item{\code{...}}{Additional parameters fed to Conos$plotGraph()
Save summary metrics to text file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveSummaryMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-saveSummaryMetrics}{}}}
\subsection{Method \code{saveSummaryMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveSummaryMetrics(
  file = "Summary_metrics.txt",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Output file (default = "Summary_metrics.txt")}

\item{\code{dec}}{How the decimals are defined (default = ".")}

\item{\code{sep}}{How the values are separated (default = "\t")
Save detailed metrics to text file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-saveDetailedMetrics}{}}}
\subsection{Method \code{saveDetailedMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveDetailedMetrics(
  file = "Detailed_metrics.txt",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Output file (default = "Summary_metrics.txt")}

\item{\code{dec}}{How the decimals are defined (default = ".")}

\item{\code{sep}}{How the values are separated (default = "\t")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
