% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRMetrics.R
\name{CRMetrics}
\alias{CRMetrics}
\title{CRMetrics class object}
\description{
CRMetrics class object

CRMetrics class object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{(default = NULL)}

\item{\code{data_path}}{(default = NULL)}

\item{\code{summary_metrics}}{(default = NULL)}

\item{\code{detailed_metrics}}{(default = NULL)}

\item{\code{comp_group}}{(default = NULL)}

\item{\code{verbose}}{(default = TRUE)}

\item{\code{pal}}{(default = NULL)}

\item{\code{theme}}{(default = NULL)}

\item{\code{n.cores}}{(default = 1)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CRMetrics$new()}}
\item \href{#method-addDetailedMetrics}{\code{CRMetrics$addDetailedMetrics()}}
\item \href{#method-plotSamples}{\code{CRMetrics$plotSamples()}}
\item \href{#method-plotGeneCounts}{\code{CRMetrics$plotGeneCounts()}}
\item \href{#method-plotUMICounts}{\code{CRMetrics$plotUMICounts()}}
\item \href{#method-plotMedianUMI}{\code{CRMetrics$plotMedianUMI()}}
\item \href{#method-plotSummaryStats}{\code{CRMetrics$plotSummaryStats()}}
\item \href{#method-plotMedianGene}{\code{CRMetrics$plotMedianGene()}}
\item \href{#method-plotCells}{\code{CRMetrics$plotCells()}}
\item \href{#method-saveSummaryMetrics}{\code{CRMetrics$saveSummaryMetrics()}}
\item \href{#method-saveDetailedMetrics}{\code{CRMetrics$saveDetailedMetrics()}}
\item \href{#method-clone}{\code{CRMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$new(
  data_path,
  metadata_file,
  comp_group = NULL,
  detailed_metrics = FALSE,
  verbose = TRUE,
  pal = "pearl_earring",
  theme = theme_bw(),
  n.cores = 1,
  version = "V2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Print messages (default = TRUE)}

\item{\code{pal}}{Palette from package 'dutchmasters' for plotting (default: "pearl_earring")}

\item{\code{theme}}{Ggplot2 theme (default: theme_bw())}

\item{\code{n.cores}}{Number of cores for calculations (default = 1)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-addDetailedMetrics}{}}}
\subsection{Method \code{addDetailedMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addDetailedMetrics(
  version = "V2",
  samples = self$metadata$sample,
  data_path = self$data_path,
  n.cores = self$n.cores,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.cores}}{Number of cores for calculations (default = 1)}

\item{\code{verbose}}{Print messages (default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSamples"></a>}}
\if{latex}{\out{\hypertarget{method-plotSamples}{}}}
\subsection{Method \code{plotSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSamples(
  pos = 2000,
  exact = FALSE,
  comp_group = self$comp_group,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneCounts"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneCounts}{}}}
\subsection{Method \code{plotGeneCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotGeneCounts(
  comp_group = self$comp_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotUMICounts"></a>}}
\if{latex}{\out{\hypertarget{method-plotUMICounts}{}}}
\subsection{Method \code{plotUMICounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUMICounts(
  comp_group = self$comp_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMedianUMI"></a>}}
\if{latex}{\out{\hypertarget{method-plotMedianUMI}{}}}
\subsection{Method \code{plotMedianUMI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotMedianUMI(
  pos = 2000,
  exact = FALSE,
  comp_group = self$comp_group,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSummaryStats"></a>}}
\if{latex}{\out{\hypertarget{method-plotSummaryStats}{}}}
\subsection{Method \code{plotSummaryStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSummaryStats(
  pos = 2000,
  exact = FALSE,
  metrics = NULL,
  comp_group = self$comp_group,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMedianGene"></a>}}
\if{latex}{\out{\hypertarget{method-plotMedianGene}{}}}
\subsection{Method \code{plotMedianGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotMedianGene(
  pos = 500,
  exact = FALSE,
  comp_group = self$comp_group,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCells"></a>}}
\if{latex}{\out{\hypertarget{method-plotCells}{}}}
\subsection{Method \code{plotCells()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotCells(
  pos = 2000,
  exact = FALSE,
  comp_group = self$comp_group,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveSummaryMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-saveSummaryMetrics}{}}}
\subsection{Method \code{saveSummaryMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveSummaryMetrics(
  file = "Summary_metrics.txt",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-saveDetailedMetrics}{}}}
\subsection{Method \code{saveDetailedMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveDetailedMetrics(
  file = "Detailed_metrics.txt",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
