% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRMetrics.R
\name{CRMetrics}
\alias{CRMetrics}
\title{CRMetrics class object}
\description{
CRMetrics class object

CRMetrics class object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{(default = NULL)}

\item{\code{data_path}}{(default = NULL)}

\item{\code{summary_metrics}}{(default = NULL)}

\item{\code{detailed_metrics}}{(default = NULL)}

\item{\code{comp_group}}{(default = NULL)}

\item{\code{verbose}}{(default = TRUE)}

\item{\code{pal}}{(default = NULL)}

\item{\code{theme}}{(default = NULL)}

\item{\code{n.cores}}{(default = 1)
Inner function to add detailed metrics}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CRMetrics-new}{\code{CRMetrics$new()}}
\item \href{#method-CRMetrics-addDetailedMetrics}{\code{CRMetrics$addDetailedMetrics()}}
\item \href{#method-CRMetrics-addComparison}{\code{CRMetrics$addComparison()}}
\item \href{#method-CRMetrics-plotSamples}{\code{CRMetrics$plotSamples()}}
\item \href{#method-CRMetrics-plotGeneCounts}{\code{CRMetrics$plotGeneCounts()}}
\item \href{#method-CRMetrics-plotUmiCounts}{\code{CRMetrics$plotUmiCounts()}}
\item \href{#method-CRMetrics-plotMedianUmi}{\code{CRMetrics$plotMedianUmi()}}
\item \href{#method-CRMetrics-plotSummaryStats}{\code{CRMetrics$plotSummaryStats()}}
\item \href{#method-CRMetrics-plotMedianGene}{\code{CRMetrics$plotMedianGene()}}
\item \href{#method-CRMetrics-plotCells}{\code{CRMetrics$plotCells()}}
\item \href{#method-CRMetrics-plotUMAP}{\code{CRMetrics$plotUMAP()}}
\item \href{#method-CRMetrics-saveSummaryMetrics}{\code{CRMetrics$saveSummaryMetrics()}}
\item \href{#method-CRMetrics-saveDetailedMetrics}{\code{CRMetrics$saveDetailedMetrics()}}
\item \href{#method-CRMetrics-clone}{\code{CRMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$new(
  data_path,
  metadata_file,
  comp_group = NULL,
  detailed_metrics = FALSE,
  verbose = TRUE,
  pal = "pearl_earring",
  theme = theme_bw(),
  n.cores = 1,
  version = "V2"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{Print messages (default = TRUE)}

\item{\code{pal}}{Palette from package 'dutchmasters' for plotting (default: "pearl_earring")}

\item{\code{theme}}{Ggplot2 theme (default: theme_bw())}

\item{\code{n.cores}}{Number of cores for calculations (default = 1)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-addDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-addDetailedMetrics}{}}}
\subsection{Method \code{addDetailedMetrics()}}{
Function to read in detailed metrics. This is not done upon initialization for speed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addDetailedMetrics(
  version = "V2",
  samples = self$metadata$sample,
  data_path = self$data_path,
  n.cores = self$n.cores,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{(default = "V2)}

\item{\code{samples}}{vector containing samples. Default is to extract this information from self$metadata$sample}

\item{\code{data_path}}{Path to data (default = self$data_path)}

\item{\code{n.cores}}{Number of cores for the calculations (default = self$n.cores)}

\item{\code{verbose}}{Print messages or not (default = self$verbose)
Add comparison group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-addComparison"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-addComparison}{}}}
\subsection{Method \code{addComparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addComparison(comp_group, metadata = self$metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric, e.g. "group", "samples"}

\item{\code{metadata}}{Metadata for samples (default = self$metadata)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector
Plot samples
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotSamples"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotSamples}{}}}
\subsection{Method \code{plotSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSamples(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric (default = self$comp_group)}

\item{\code{h.adj}}{Position of statistics test p value as \% of max(y) (default = 0.05)}

\item{\code{exact}}{Whether to calculate exact p values (default = FALSE)}

\item{\code{metadata}}{Metadata for samples (default = self$metadata)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
Plot gene counts of raw data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotGeneCounts"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotGeneCounts}{}}}
\subsection{Method \code{plotGeneCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotGeneCounts(
  comp_group = self$comp_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Plot UMI counts of raw data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotUmiCounts"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotUmiCounts}{}}}
\subsection{Method \code{plotUmiCounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUmiCounts(
  comp_group = self$comp_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Plot median UMI counts}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotMedianUmi"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotMedianUmi}{}}}
\subsection{Method \code{plotMedianUmi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotMedianUmi(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Plot all summary stats or a selected list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotSummaryStats"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotSummaryStats}{}}}
\subsection{Method \code{plotSummaryStats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSummaryStats(
  comp_group = self$comp_group,
  metrics = NULL,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Summary stat plot for median gene number per cell}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotMedianGene"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotMedianGene}{}}}
\subsection{Method \code{plotMedianGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotMedianGene(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Plot cells}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotCells"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotCells}{}}}
\subsection{Method \code{plotCells()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotCells(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_metrics}}{Plot cells in a UMAP using Conos and color by depth and doublets}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-plotUMAP"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-plotUMAP}{}}}
\subsection{Method \code{plotUMAP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUMAP(
  cms = self$cm.list,
  preprocess = "Pagoda2",
  ncores = self$n.cores,
  cutoff_depth = 10000,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cms}}{The count matrices (default = self$cm.list)}

\item{\code{preprocess}}{The method used for preprocessing, either Pagoda2 or Seurat (default = Pagoda2)}

\item{\code{ncores}}{The number of cores to use (default = self$n.cores)}

\item{\code{cutoff_depth}}{The depth cutoff to color the UMAP (default = 1e4)
Save summary metrics to text file}

\item{\code{verbose}}{Print messages (default = TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-saveSummaryMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-saveSummaryMetrics}{}}}
\subsection{Method \code{saveSummaryMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveSummaryMetrics(
  file = "Summary_metrics.txt",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{Save detailed metrics to text file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-saveDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-saveDetailedMetrics}{}}}
\subsection{Method \code{saveDetailedMetrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveDetailedMetrics(
  file = "Detailed_metrics.txt",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CRMetrics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CRMetrics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
