% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRMetrics.R
\name{CRMetrics}
\alias{CRMetrics}
\title{CRMetrics class object}
\description{
Functions to analyse cellranger count data.
}
\examples{

## ------------------------------------------------
## Method `CRMetrics$new`
## ------------------------------------------------

crm <- CRMetrics$new(data_path = "data/CRMetrics_testdata")
Add detailed metrics

## ------------------------------------------------
## Method `CRMetrics$addDetailedMetrics`
## ------------------------------------------------

crm$addDetailedMetrics()
Add comparison group

## ------------------------------------------------
## Method `CRMetrics$addComparison`
## ------------------------------------------------

crm$addComparison(comp_group = "sex")
Plot samples

## ------------------------------------------------
## Method `CRMetrics$plotSamples`
## ------------------------------------------------

crm$plotSamples(comp_group = "sex", second_comp_group = "condition")
Plot summary metrics

## ------------------------------------------------
## Method `CRMetrics$plotSummaryMetrics`
## ------------------------------------------------

metrics.to.plot <- crm$selectMetrics(ids = c(1:4, 6, 18, 19)) 
crm$plotSummaryMetrics(metrics = metrics.to.plot, plot_geom = "point")
Plot detailed metrics

## ------------------------------------------------
## Method `CRMetrics$plotDetailedMetrics`
## ------------------------------------------------

metrics.to.plot <- crm$detailed_metrics$metric \%>\% unique()
crm$plotDetailedMetrics()
Plot UMAP

## ------------------------------------------------
## Method `CRMetrics$plotUmap`
## ------------------------------------------------

crm$doPreprocessing()
crm$createEmbedding() 
crm$plotUMAP()
# Color cells for low depth
crm$plotUMAP(depth = TRUE, depth.cutoff = 1e3)
Plot depth

## ------------------------------------------------
## Method `CRMetrics$plotDepth`
## ------------------------------------------------

crm$plotDepth()
Save summary metrics

## ------------------------------------------------
## Method `CRMetrics$saveSummaryMetrics`
## ------------------------------------------------

crm$saveSummaryMetrics(file = "Summary_metrics.tsv")
Save detailed metrics

## ------------------------------------------------
## Method `CRMetrics$saveDetailedMetrics`
## ------------------------------------------------

crm$saveDetailedMetrics(file = "Detailed_metrics.tsv")
Detect doublets

## ------------------------------------------------
## Method `CRMetrics$detectDoublets`
## ------------------------------------------------

crm$detectDoublets(method = "scrublet", conda.path = "/opt/software/miniconda/4.12.0/condabin/conda")
Conos preprocessing

## ------------------------------------------------
## Method `CRMetrics$doPreprocessing`
## ------------------------------------------------

crm$doPreprocessing(preprocess = "pagoda2")
Create Conos embedding

## ------------------------------------------------
## Method `CRMetrics$createEmbedding`
## ------------------------------------------------

crm$createEmbedding()
Filter count matrices

## ------------------------------------------------
## Method `CRMetrics$filterCms`
## ------------------------------------------------

crm$filterCMS(file = "cms_filtered.rds", depth_cutoff = 1e3, mito_cutoff = 0.05, doublets = "scrublet")
Select summary metrics.

## ------------------------------------------------
## Method `CRMetrics$selectMetrics`
## ------------------------------------------------

crm$selectMetrics()
selection.metrics <- crm$selectMetrics(c(1:4))
Plot filtered cells

## ------------------------------------------------
## Method `CRMetrics$plotFilteredCells`
## ------------------------------------------------

crm$plotFilteredCells(type = "umap", doublet_method = "scrublet")
filtered.cells <- crm$plotFilteredCells(type = "export", doublet_method = "scrublet")
Get Conos sequencing depth

## ------------------------------------------------
## Method `CRMetrics$getConosDepth`
## ------------------------------------------------

crm$getConosDepth()
Get fraction of mitochondrial genes

## ------------------------------------------------
## Method `CRMetrics$getMitoFraction`
## ------------------------------------------------

crm$getMitoFraction(species = c("human", "mouse"))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{metadata}}{(default = NULL)}

\item{\code{data_path}}{(default = NULL)}

\item{\code{summary_metrics}}{(default = NULL)}

\item{\code{detailed_metrics}}{(default = NULL)}

\item{\code{comp_group}}{(default = NULL)}

\item{\code{verbose}}{(default = TRUE)}

\item{\code{theme}}{(default = NULL)}

\item{\code{n.cores}}{(default = 1)
Initialize a CRMetrics object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CRMetrics$new()}}
\item \href{#method-addDetailedMetrics}{\code{CRMetrics$addDetailedMetrics()}}
\item \href{#method-addComparison}{\code{CRMetrics$addComparison()}}
\item \href{#method-plotSamples}{\code{CRMetrics$plotSamples()}}
\item \href{#method-plotSummaryMetrics}{\code{CRMetrics$plotSummaryMetrics()}}
\item \href{#method-plotDetailedMetrics}{\code{CRMetrics$plotDetailedMetrics()}}
\item \href{#method-plotUmap}{\code{CRMetrics$plotUmap()}}
\item \href{#method-plotDepth}{\code{CRMetrics$plotDepth()}}
\item \href{#method-saveSummaryMetrics}{\code{CRMetrics$saveSummaryMetrics()}}
\item \href{#method-saveDetailedMetrics}{\code{CRMetrics$saveDetailedMetrics()}}
\item \href{#method-detectDoublets}{\code{CRMetrics$detectDoublets()}}
\item \href{#method-doPreprocessing}{\code{CRMetrics$doPreprocessing()}}
\item \href{#method-createEmbedding}{\code{CRMetrics$createEmbedding()}}
\item \href{#method-filterCms}{\code{CRMetrics$filterCms()}}
\item \href{#method-selectMetrics}{\code{CRMetrics$selectMetrics()}}
\item \href{#method-plotFilteredCells}{\code{CRMetrics$plotFilteredCells()}}
\item \href{#method-getConosDepth}{\code{CRMetrics$getConosDepth()}}
\item \href{#method-getMitoFraction}{\code{CRMetrics$getMitoFraction()}}
\item \href{#method-clone}{\code{CRMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
To initialize new object, data_path is needed. metadata_file is also recommended, but not required.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$new(
  data_path,
  metadata = NULL,
  comp_group = NULL,
  detailed_metrics = FALSE,
  verbose = TRUE,
  theme = theme_bw(),
  n.cores = 1,
  raw.meta = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_path}}{Path to cellranger count data (default = NULL).}

\item{\code{metadata}}{Path to metadata file or name of metadata object (default = NULL).}

\item{\code{comp_group}}{A group present in the metadata to compare the metrics by, can be added with addComparison (default = NULL).}

\item{\code{detailed_metrics}}{Object containing a data frame with the detailed metrics, can be added with addDetailedMetrics (default = NULL).}

\item{\code{verbose}}{Print messages or not (default = TRUE).}

\item{\code{theme}}{Ggplot2 theme (default: theme_bw()).}

\item{\code{n.cores}}{Number of cores for the calculations (default = self$n.cores).}

\item{\code{raw.meta}}{Keep metadata in its raw format. If FALSE, classes will be converted using "type.convert" (default = FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CRMetrics object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm <- CRMetrics$new(data_path = "data/CRMetrics_testdata")
Add detailed metrics
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-addDetailedMetrics}{}}}
\subsection{Method \code{addDetailedMetrics()}}{
Function to read in detailed metrics. This is not done upon initialization for speed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addDetailedMetrics(
  data_path = self$data_path,
  samples = self$metadata$sample,
  transcript = "SYMBOL",
  sep = "!!",
  n.cores = self$n.cores,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_path}}{Path to cellranger count data (default = self$data_path).}

\item{\code{samples}}{Vector containing samples (default = self$metadata$sample).}

\item{\code{transcript}}{The type of transcript, SYMBOL or ENSEMBLE (default = "SYMBOL").}

\item{\code{sep}}{Separator for cell names (default = "!!").}

\item{\code{n.cores}}{Number of cores for the calculations (default = self$n.cores).}

\item{\code{verbose}}{Print messages or not (default = self$verbose).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Count matrices
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$addDetailedMetrics()
Add comparison group
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addComparison"></a>}}
\if{latex}{\out{\hypertarget{method-addComparison}{}}}
\subsection{Method \code{addComparison()}}{
Add comparison group for statistical testing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$addComparison(comp_group, metadata = self$metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric (default = self$comp_group).}

\item{\code{metadata}}{Metadata for samples (default = self$metadata).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$addComparison(comp_group = "sex")
Plot samples
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSamples"></a>}}
\if{latex}{\out{\hypertarget{method-plotSamples}{}}}
\subsection{Method \code{plotSamples()}}{
Plot the number of samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSamples(
  comp_group = self$comp_group,
  h.adj = 0.05,
  exact = FALSE,
  metadata = self$metadata,
  second_comp_group = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric (default = self$comp_group).}

\item{\code{h.adj}}{Position of statistics test p value as \% of max(y) (default = 0.05).}

\item{\code{exact}}{Whether to calculate exact p values (default = FALSE).}

\item{\code{metadata}}{Metadata for samples (default = self$metadata).}

\item{\code{second_comp_group}}{Second comparison metric (default = NULL).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$plotSamples(comp_group = "sex", second_comp_group = "condition")
Plot summary metrics
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSummaryMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-plotSummaryMetrics}{}}}
\subsection{Method \code{plotSummaryMetrics()}}{
Plot all summary stats or a selected list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotSummaryMetrics(
  comp_group = self$comp_group,
  second_comp_group = NULL,
  metrics = NULL,
  h.adj = 0.05,
  plot_stat = TRUE,
  stat_test = c("non_parametric", "parametric"),
  exact = FALSE,
  metadata = self$metadata,
  summary_metrics = self$summary_metrics,
  plot_geom = "point",
  se = FALSE,
  group_reg_lines = FALSE,
  secondary_testing = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric (default = self$comp_group).}

\item{\code{second_comp_group}}{Second comparison metric, used for the metric "samples per group" or when "comp_group" is a numeric or an integer (default = NULL).}

\item{\code{metrics}}{Metrics to plot (default = NULL).}

\item{\code{h.adj}}{Position of statistics test p value as \% of max(y) (default = 0.05)}

\item{\code{plot_stat}}{Show statistics in plot. Will be FALSE if "comp_group" = "sample" or if "comp_group" is a numeric or an integer (default = TRUE)}

\item{\code{stat_test}}{Statistical test to perform to compare means. Can either be "non_parametric" or "parametric" (default = "non_parametric").}

\item{\code{exact}}{Whether to calculate exact p values (default = FALSE).}

\item{\code{metadata}}{Metadata for samples (default = self$metadata).}

\item{\code{summary_metrics}}{Summary metrics (default = self$summary_metrics).}

\item{\code{plot_geom}}{Which geometric is used to plot the data (default = "point").}

\item{\code{se}}{For regression lines, show SE (default = FALSE)}

\item{\code{group_reg_lines}}{For regression lines, if FALSE show one line, if TRUE show line per group defined by second_comp_group (default = FALSE)}

\item{\code{secondary_testing}}{Whether to show post hoc testing (default = TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{metrics.to.plot <- crm$selectMetrics(ids = c(1:4, 6, 18, 19)) 
crm$plotSummaryMetrics(metrics = metrics.to.plot, plot_geom = "point")
Plot detailed metrics
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-plotDetailedMetrics}{}}}
\subsection{Method \code{plotDetailedMetrics()}}{
Plot detailed metrics from the detailed_metrics object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotDetailedMetrics(
  comp_group = self$com_group,
  detailed_metrics = self$detailed_metrics,
  metadata = self$metadata,
  metrics = NULL,
  plot_geom = NULL,
  data_path = self$data_path,
  hline = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp_group}}{Comparison metric (default = self$comp_group).}

\item{\code{detailed_metrics}}{Object containing the count matrices (default = self$detailed_metrics).}

\item{\code{metadata}}{Metadata for samples (default = self$metadata).}

\item{\code{metrics}}{Metrics to plot (default = NULL).}

\item{\code{plot_geom}}{How to plot the data (default = NULL).}

\item{\code{data_path}}{Path to cellranger count data (default = self$data_path).}

\item{\code{hline}}{Whether to show median as horizontal line (default = TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{metrics.to.plot <- crm$detailed_metrics$metric \%>\% unique()
crm$plotDetailedMetrics()
Plot UMAP
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotUmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotUmap}{}}}
\subsection{Method \code{plotUmap()}}{
Plot cells in a UMAP using Conos and color by depth and doublets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotUmap(
  depth = FALSE,
  doublet_method = NULL,
  doublet_scores = FALSE,
  depth.cutoff = 1000,
  mito.frac = FALSE,
  mito.cutoff = 0.05,
  species = c("human", "mouse"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{Plot depth or not (default = FALSE).}

\item{\code{doublet_method}}{Doublet detection method (default = NULL).}

\item{\code{doublet_scores}}{Plot doublet scores or not (default = FALSE).}

\item{\code{depth.cutoff}}{Depth cutoff (default = 1e33).}

\item{\code{mito.frac}}{Plot mitochondrial fraction or not (default = FALSE).}

\item{\code{mito.cutoff}}{Mitochondrial fraction cutoff (default = 0.05).}

\item{\code{species}}{Species to calculate the mitochondrial fraction for (default = c("human","mouse")).}

\item{\code{...}}{Plotting parameters passed to \code{conos::embeddingPlot}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$doPreprocessing()
crm$createEmbedding() 
crm$plotUMAP()
# Color cells for low depth
crm$plotUMAP(depth = TRUE, depth.cutoff = 1e3)
Plot depth
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDepth"></a>}}
\if{latex}{\out{\hypertarget{method-plotDepth}{}}}
\subsection{Method \code{plotDepth()}}{
Plot the sequencing depth in histogram.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotDepth(cutoff = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cutoff}}{The depth cutoff to color the UMAP (default = 1e3).}

\item{\code{per_sample}}{Whether to plot the depth per sample or for all the samples (default = FALSE).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$plotDepth()
Save summary metrics
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveSummaryMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-saveSummaryMetrics}{}}}
\subsection{Method \code{saveSummaryMetrics()}}{
Save summary metrics to text file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveSummaryMetrics(
  file = "Summary_metrics.tsv",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Output file (default = "Summary_metrics.txt").}

\item{\code{dec}}{How the decimals are defined (default = ".").}

\item{\code{sep}}{What separator to use (default = "\t").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
file
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$saveSummaryMetrics(file = "Summary_metrics.tsv")
Save detailed metrics
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDetailedMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-saveDetailedMetrics}{}}}
\subsection{Method \code{saveDetailedMetrics()}}{
Save detailed metrics to text file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$saveDetailedMetrics(
  file = "Detailed_metrics.tsv",
  dec = ".",
  sep = "\\t"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Output file (default = "Summary_metrics.tsv").}

\item{\code{dec}}{How the decimals are defined (default = ".").}

\item{\code{sep}}{What separator to use (default = "\t").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
file
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$saveDetailedMetrics(file = "Detailed_metrics.tsv")
Detect doublets
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-detectDoublets"></a>}}
\if{latex}{\out{\hypertarget{method-detectDoublets}{}}}
\subsection{Method \code{detectDoublets()}}{
Detect doublet cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$detectDoublets(
  method = c("scrublet", "doubletdetection"),
  cms = self$cms,
  env = "r-reticulate",
  conda.path = system("whereis conda"),
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Which method to use, either \code{scrublet} or \code{doubletdetection} (default="scrublet").}

\item{\code{cms}}{List containing the count matrices (default=self$cms).}

\item{\code{env}}{Environment to run python in (default="r-reticulate").}

\item{\code{conda.path}}{Path to conda environment (default=system("whereis conda")).}

\item{\code{verbose}}{Print messages or not (defeults = self$verbose).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$detectDoublets(method = "scrublet", conda.path = "/opt/software/miniconda/4.12.0/condabin/conda")
Conos preprocessing
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doPreprocessing"></a>}}
\if{latex}{\out{\hypertarget{method-doPreprocessing}{}}}
\subsection{Method \code{doPreprocessing()}}{
Perform conos preprocessing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$doPreprocessing(
  cms = self$cms,
  preprocess = c("pagoda2", "seurat"),
  verbose = self$verbose,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cms}}{List containing the count matrices (default = self$cms).}

\item{\code{preprocess}}{Method to use for preprocessing (default = c("pagoda2","seurat")).}

\item{\code{verbose}}{Print messages or not (default = self$verbose).}

\item{\code{n.cores}}{Number of cores for the calculations (default = self$n.cores).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conos object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$doPreprocessing(preprocess = "pagoda2")
Create Conos embedding
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-createEmbedding}{}}}
\subsection{Method \code{createEmbedding()}}{
Create conos UMAP embedding.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$createEmbedding(
  cms = self$cms.preprocessed,
  verbose = self$verbose,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cms}}{List containing the count matrices (default = self$cms).}

\item{\code{verbose}}{Print messages or not (default = self$verbose).}

\item{\code{n.cores}}{Number of cores for the calculations (default = self$n.cores).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Conos object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$createEmbedding()
Filter count matrices
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filterCms"></a>}}
\if{latex}{\out{\hypertarget{method-filterCms}{}}}
\subsection{Method \code{filterCms()}}{
Filter cells based on depth, mitochondrial fraction and doublets from the count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$filterCms(
  file = "cms_filtered.rds",
  raw = TRUE,
  depth_cutoff = NULL,
  mito_cutoff = NULL,
  doublets = NULL,
  compress = FALSE,
  species = c("human", "mouse"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{File to save filtered count matrices to (default = "cms_filtered.rds").}

\item{\code{raw}}{Save raw count matrices (TRUE) or normalized count matrices (default = TRUE).}

\item{\code{depth_cutoff}}{Depth cutoff (default = NULL).}

\item{\code{mito_cutoff}}{Mitochondrial fraction cutoff (default = NULL).}

\item{\code{doublets}}{Doublet detection method to use (default = NULL).}

\item{\code{compress}}{Compress the file or not (default = FALSE).}

\item{\code{species}}{Species to calculate the mitochondrial fraction for (default = "human").}

\item{\code{...}}{Parameters for saving R object passed to \code{saveRDS}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
file
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$filterCMS(file = "cms_filtered.rds", depth_cutoff = 1e3, mito_cutoff = 0.05, doublets = "scrublet")
Select summary metrics.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-selectMetrics}{}}}
\subsection{Method \code{selectMetrics()}}{
Select metrics from summary_metrics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$selectMetrics(ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Metric id to select (default = NULL).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$selectMetrics()
selection.metrics <- crm$selectMetrics(c(1:4))
Plot filtered cells
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFilteredCells"></a>}}
\if{latex}{\out{\hypertarget{method-plotFilteredCells}{}}}
\subsection{Method \code{plotFilteredCells()}}{
Plot filetered cells on a UMAP, in a bar plot, on a tile or export the data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$plotFilteredCells(
  type = c("umap", "bar", "tile", "export"),
  depth = TRUE,
  depth.cutoff = 1000,
  doublet_method = NULL,
  mito.frac = TRUE,
  mito.cutoff = 0.05,
  species = c("human", "mouse")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of plot to use: umap, bar, tile or export (default = c("umap","bar","tile","export")).}

\item{\code{depth}}{Plot the depth or not (default = TRUE).}

\item{\code{depth.cutoff}}{Depth cutoff (default = 1e3).}

\item{\code{doublet_method}}{Method to detect doublets (default = NULL).}

\item{\code{mito.frac}}{Plot the mitochondrial fraction or not (default = TRUE).}

\item{\code{mito.cutoff}}{Mitochondrial fraction cutoff (default = 0.05).}

\item{\code{species}}{Species to calculate the mitochondrial fraction for (default = c("human","mouse")).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object or data frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$plotFilteredCells(type = "umap", doublet_method = "scrublet")
filtered.cells <- crm$plotFilteredCells(type = "export", doublet_method = "scrublet")
Get Conos sequencing depth
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConosDepth"></a>}}
\if{latex}{\out{\hypertarget{method-getConosDepth}{}}}
\subsection{Method \code{getConosDepth()}}{
Extract sequencing depth from Conos object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$getConosDepth()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$getConosDepth()
Get fraction of mitochondrial genes
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMitoFraction"></a>}}
\if{latex}{\out{\hypertarget{method-getMitoFraction}{}}}
\subsection{Method \code{getMitoFraction()}}{
Calculate the fraction of mitochondrial genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$getMitoFraction(species = "human")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species}}{Species to calculate the mitochondrial fraction for (default = "human").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data frame
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{crm$getMitoFraction(species = c("human", "mouse"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CRMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
