% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{filterVector}
\alias{filterVector}
\title{Create filtering vector}
\usage{
filterVector(num.vec, name, filter, samples, sep = "!!")
}
\arguments{
\item{num.vec}{numeric Numeric vector to create filter on}

\item{name}{character Name of filter}

\item{filter}{numeric Either a single numeric value or a numeric value with length of samples}

\item{samples}{character Sample IDs}

\item{sep}{character Separator to split cells by into sample-wise lists (default = "!!")}
}
\description{
Create logical filtering vector based on a numeric vector and a (sample-wise) cutoff
}
\keyword{internal}
