% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{labelsFilter}
\alias{labelsFilter}
\title{Get labels for percentage of filtered cells}
\usage{
labelsFilter(filter.data)
}
\arguments{
\item{filter.data}{Data frame containing the mitochondrial fraction, depth and doublets per sample.}
}
\value{
data frame
}
\description{
Labels the percentage of filtered cells based on mitochondrial fraction, sequencing depth and doublets as low, medium or high
}
\examples{
\dontrun{
filtered <- labelsFilter(filter.data)
}
}
\keyword{internal}
