% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{plotGeom}
\alias{plotGeom}
\title{Plot the data as points, as bars or as a histogram}
\usage{
plotGeom(plot_geom)
}
\arguments{
\item{plot_geom}{The plot_geom to use, "point", "bar", "histogram", or "violin"}
}
\description{
Plot the data as points, as bars or as a histogram
}
\keyword{internal}
