% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{read10xH5}
\alias{read10xH5}
\title{Read 10x HDF5 files}
\usage{
read10xH5(
  data.path,
  sample.names = NULL,
  type = c("raw", "filtered", "cellbender", "cellbender_filtered"),
  symbol = TRUE,
  sep = "!!",
  n.cores = 1,
  verbose = TRUE,
  unique.names = FALSE
)
}
\arguments{
\item{sample.names}{character vector, select specific samples for processing (default = NULL)}

\item{type}{name of H5 file to search for, "raw" and "filtered" are Cell Ranger count outputs, "cellbender" is output from CellBender after running script from saveCellbenderScript}
}
\description{
Read 10x HDF5 files
}
